Move result chip to legacy file
